requirements.txt
streamlit
pandas
pydeck
openpyxl
requests
plotly
pillow
torch
torchvision
transformers
numpy